requirements.txt
# Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Data Processing
pandas==2.1.4
python-dotenv==1.0.0

# API (if using FastAPI)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Utilities
httpx==0.25.2
tenacity==8.2.3
fake-useragent==1.4.0

# Development
pytest==7.4.3
black==23.11.0
pylint==3.0.2